+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed a small syntax highlighting issue.
 - Fixed incorrect function call parsing.
+- Fixed double function call parsing.
 
 ## [2.3.1]